Change the plan: use streaming for upload, separate step for key reset
There is no temporary file created anymore for the temporary archive,
the archive is streamed to the remote end through SSH instead.

Resetting the keys requires a back and forth between the server
and the client, to connect to run the reset, then forget previous
keys locally, then connect again to check the new keys.