Do not download pam_ssh_agent_agent from SourceForge anymore
The sources are included as a git submodule and uploaded to the
server part of the tarball of setup scripts.

As a result, the version of the library to deploy is no longer
set in the install script, but by selecting the corresponding
commit in the git submodule sudo/pam_ssh_agent_auth.

Related issue: #3